refactor: nest reader execution modules
